refactor (remove AuditLogClient interface)
