--- src/EventListener/AuditLogListener/AuditLogClient.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Landingi\EventStoreBundle\EventListener\AuditLogListener;
-
-interface AuditLogClient
-{
-    public function store(AuditLogEvent $event): void;
-}
